Update commit from code review
Updated the commit for issue #93 based on code review.

+++ apps/app_rpt/rpt_config.h
@@ -24,7 +24,7 @@ int elink_db_get(char *lookup, char c, char *nodenum, char *callsign, char *ipad
 int tlb_node_get(char *lookup, char c, char *nodenum, char *callsign, char *ipaddr, char *port);
 
 /*!
- * \brief AllStar Network node lookup function.  This function will take the nodelist that has been read into memory
+ * \brief Node lookup function.  This function will take the nodelist that has been read into memory
  * and try to match the node number that was passed to it.  If it is found, the function requested will succeed.
  * If not, it will fail.  Called when a connection to a remote node is requested.
  * \param  myrpt		Calling repeater structure
@@ -32,24 +32,24 @@ int tlb_node_get(char *lookup, char c, char *nodenum, char *callsign, char *ipad
  * \param  nodedata		A buffer to hold the matching node information
  * \param  nodedatalength	The length of the str buffer
  * \param  wilds		Set to 1 to perform a wild card lookup
- * \retval 0 			If not successful
- * \retval 1 			If successful
+ * \retval -1 			If not successful
+ * \retval 0 			If successful
  */
-int node_lookup(struct rpt *myrpt, char *digitbuf, char *nodedata, int nodedatalength, int wilds);
+int node_lookup(struct rpt *myrpt, char *digitbuf, char *nodedata, size_t nodedatalength, int wilds);
 
 /*!
- * \brief AllStar Network forwarde node lookup function.  This function will take the nodelist 
+ * \brief Forward node lookup function.  This function will take the nodelist 
  * and try to match the node number that was passed to it.  If it is found, the function requested will succeed.
  * If not, it will fail.  Called when a connection to a remote node is requested.
  * \param  digitbuf		The node number of match
  * \param  cfg			Asterisk configuration file pointer
  * \param  nodedata		A buffer to hold the matching node information
  * \param  nodedatalength	The length of the str buffer
- * \retval 0 			If not successful
- * \retval 1 			If successful
+ * \retval -1 			If not successful
+ * \retval 0 			If successful
  */
 
-int forward_node_lookup(char *digitbuf, struct ast_config *cfg, char *nodedata, int nodedatalength);
+int forward_node_lookup(char *digitbuf, struct ast_config *cfg, char *nodedata, size_t nodedatalength);
 
 /*! 
  * \brief This is the initialization function.  This routine takes the data in rpt.conf and setup up the variables needed for each of
